+++ src/main/kotlin/net/plpgsql/ideadebugger/Diagnostic.kt
@@ -9,7 +9,9 @@ import com.intellij.openapi.project.Project
 import net.plpgsql.ideadebugger.command.PlActivity
 import net.plpgsql.ideadebugger.ui.ConnectionDiagnosticDialog
 
-data class ExtensionDiagnostic (
+data class ConnectionDiagnostic (
+    var customCommandMessage: String,
+    var customCommandOk: Boolean,
     var sharedLibraries: String,
     var sharedLibraryOk: Boolean,
     var extensions: String,
@@ -18,10 +20,10 @@ data class ExtensionDiagnostic (
     var activityOk: Boolean
 )
 
-fun extensionOk(diag: ExtensionDiagnostic): Boolean =
-    diag.extensionOk && diag.sharedLibraryOk && diag.activityOk
+fun extensionOk(diag: ConnectionDiagnostic): Boolean =
+    diag.customCommandOk && diag.extensionOk && diag.sharedLibraryOk && diag.activityOk
 
-fun showExtensionDiagnostic(project: Project, diag: ExtensionDiagnostic) {
+fun showExtensionDiagnostic(project: Project, diag: ConnectionDiagnostic) {
     val dialog = ConnectionDiagnosticDialog(project, diag)
     runInEdt {
         dialog.show()