+++ src/main/kotlin/net/plpgsql/ideadebugger/ui/ConnectionDiagnosticDialog.kt
@@ -9,11 +9,10 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogPanel
 import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.ui.dsl.builder.panel
-import net.plpgsql.ideadebugger.ExtensionDiagnostic
+import net.plpgsql.ideadebugger.ConnectionDiagnostic
 import javax.swing.JComponent
 
-@Suppress("UnstableApiUsage")
-class ConnectionDiagnosticDialog(project: Project, private var diagnostic: ExtensionDiagnostic) :
+class ConnectionDiagnosticDialog(project: Project, private var diagnostic: ConnectionDiagnostic) :
     DialogWrapper(project, true) {
 
     private var panel: DialogPanel? = null
@@ -27,6 +26,16 @@ class ConnectionDiagnosticDialog(project: Project, private var diagnostic: Exten
     override fun createCenterPanel(): JComponent? {
         panel = panel {
             group(title = "Plugin Diagnostic") {
+                row("Custom command") {
+                    label(diagnostic.customCommandMessage)
+                    icon(if (diagnostic.customCommandOk) AllIcons.General.InspectionsOK else AllIcons.General.Error)
+                }
+                row {
+                }.comment(
+                    """
+                    The custom command from plugin configuration
+                    """.trimIndent()
+                )
                 row("Shared libraries") {
                     label(diagnostic.sharedLibraries)
                     icon(if (diagnostic.sharedLibraryOk) AllIcons.General.InspectionsOK else AllIcons.General.Error)
@@ -34,7 +43,7 @@ class ConnectionDiagnosticDialog(project: Project, private var diagnostic: Exten
                 row {
                 }.comment(
                     """
-                    The library plugin_debugger activated in the database server 
+                    The library plugin_debugger detected in the database server 
                     Shared libraries are set in postgresql.conf
                     """.trimIndent()
                 )
@@ -42,7 +51,7 @@ class ConnectionDiagnosticDialog(project: Project, private var diagnostic: Exten
                     label(diagnostic.extensions)
                     icon(if (diagnostic.extensionOk) AllIcons.General.InspectionsOK else AllIcons.General.Error)
                 }
-                row() {
+                row {
                 }.comment(
                     """
                     Extension must be created in the current database
@@ -59,7 +68,7 @@ class ConnectionDiagnosticDialog(project: Project, private var diagnostic: Exten
                         label("Remaining session $pids")
                         icon(AllIcons.General.Error)
                     }
-                    row() {
+                    row {
                     }.comment(
                         """
                         A died debugger session can't be stopped, server you be restarted or pid killed manually